Build system.img with ota file
* You'll need an android phone ..

+++ scripts/build.sh
@@ -93,21 +93,25 @@ if [ ! -d system ]; then
   fi
 
   if [ ! -f system.img.ext4 ]; then
-    echo "Extracting stock ROM .. "
-    unzip -q $STOCK_ROM -d $UNZIPPED_STOCK_ROM_DIR
+    if [ -f $STOCK_OTA ]; then
+      $SCRIPTDIR/apply_ota_on_phone.sh
+    else 
+      echo "Extracting stock ROM .. "
+      unzip -q $STOCK_ROM -d $UNZIPPED_STOCK_ROM_DIR
 
-    echo "Move out system images .. "
-    mv $UNZIPPED_STOCK_ROM_DIR/system.* .
-    
-    move_out_image boot
+      echo "Move out system images .. "
+      mv $UNZIPPED_STOCK_ROM_DIR/system.* .
+      
+      move_out_image boot
 
-    echo "Clean up .. "
-    rm -rf $UNZIPPED_STOCK_ROM_DIR
+      echo "Clean up .. "
+      rm -rf $UNZIPPED_STOCK_ROM_DIR
 
-    echo "Converting system.new.dat to raw ext4 image .. "
-    $SCRIPTDIR/sdat2img.py system.transfer.list system.new.dat system.img.ext4
-    rm -f system.transfer.list system.new.dat system.patch.dat
-  else
+      echo "Converting system.new.dat to raw ext4 image .. "
+      $SCRIPTDIR/sdat2img.py system.transfer.list system.new.dat system.img.ext4
+      rm -f system.transfer.list system.new.dat system.patch.dat
+    fi
+  elif [ ! -f boot.img ]; then
     echo "Move out stock boot.img .. "
     unzip -q $STOCK_ROM boot.img
   fi
